hantro: add decoder media device udev rule and hantro
The use of the request API with the decoder requires us to open the
decoder media device in order to create and submit requests. However,
media devices have the same shortcomings as video devices, in that they
are just sequentially named media0, media1, ... without any indication
about which is which.

This CL identifies the decoder media devices and creates a media-dec0
symlink to it, similarly to the way we already create a video-dec0
symlink to the video decoder device. That way Chromium can open the
correct media device without any hassle.

This CL also adds the support for Hantro which is the new name for rk3288.

BUG=chromium:965375
TEST=Checked that /dev/media-dec0,video-dec,video-dec0 were visible in 4.19
  and checked /dev/video-dec,video-dec0 were visible in 3.14 upon boot.

Change-Id: Ie6a711b8230f6a99031c87eb9f0a6bfb0b321ccb
Signed-off-by: Francois Buergisser <fbuergisser@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1762093
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>