config: Add firmware-signing template to the public config
This is in the private config but does not seem to work. The private and
public configs seems to be independently processed.

We can't actually sign a public build, but we have to have this here
anyway.

BUG=chromium:993952
TEST=manual, with other CL for private config:
emerge-grunt -q chromeos-base/chromeos-config-bsp-grunt \
	chromeos-base/chromeos-config-bsp-grunt-private chromeos-config
Check in chroot /build/grunt/usr/share/chromeos-config/yaml/config.yml
for 'key-id' properties, which should be CAREENA, etc. not DEFAULT
(the config.json file only has a subset of the required info)

Change-Id: Ia3ee4196469470617238033475b9481298ebb5ba
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1808061
Reviewed-by: Martin Roth <martinroth@chromium.org>