hatch: automate adding a new variant to model.yaml
As part of creating a new variant, we need to define the device
in model.yaml and add an unprovisioned SKU (255). Because the YAML
file is pretty simple, we can just append to the end.

BUG=b:140261109
BRANCH=none
TEST=``./add_variant.sh sushi && git show``
The ebuild file will have a revbump.
files/model.yaml will have a new device-name section for 'sushi'

Change-Id: I010f1202e0fd435a24308f6fbdba8729ea39d59a
Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1810633
Reviewed-by: Jett Rink <jettrink@chromium.org>