automation: move add_variant to dev/contrib
Most of the scripts to automate creating a new variant are moving
to platform/dev/contrib/variant for maintenance reasons.
Keeping them in one place makes it easier to coordinate changes
in the scripts, instead of having them spread across different
repos. It also allows keeping documentation close to the scripts.

BUG=b:140261109
BRANCH=None
TEST=N/A, script is being removed

Change-Id: Idf14b744041f01aed682f7e74877c6dea6fe41a0
Signed-off-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1829399
Reviewed-by: Jett Rink <jettrink@chromium.org>