lakitu: systemd: Update default.target in pkg_postinst
The way we update it today has a chicken-and-egg problem. We call
systemd_getsystemunit, an eclass function, to get the system units path.
The eclass function internally calls pkg-config to resolve the
systemdsystemunit variable, which relies on the existence of
<sysroot>/usr/share/pkgconfig/systemd.pc. The latter is not installed
until src_install() is finished. So we have to do the default.target
update in one of the post install functions, e.g., pkg_postinstall().

The problem was not surfaced before because the systemd_getsystemunit
eclass function has a fallback path, which used to match our system
units path (/usr/lib/systemd/system/) but not any more since CL:1881421.

BUG=b:143545128
TEST=lakitu-release passes again
RELEASE_NOTE=None

Change-Id: I6afb1e83cf645d66a4be93eb369bebd98f9cfdbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1888915
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Commit-Queue: Daniel Wang <wonderfly@google.com>