wilco: Add --noadjtime to wilco_sync_ec_rtc.conf
Before, this script was working fine functionally, but it would
return a failure error code when you tried to stop the stop the
upstart job. This was because the last, nonessential step of
hwclock is to adjust the /etc/adjtime cache, but that is mounted
on the rootfs, which is read-only. This cleans things up so there
are not bogus error messages.

BUG=None
TEST=Ran this on my DUT, the EC RTC is still adjusted as before,
but the upstart script now returns a 0 exit code.

Change-Id: I406ef6f3b8b92df15f2ef400f7ef81169f4dc11b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1888923
Commit-Queue: Nick Crews (Nov 1 last day) <ncrews@chromium.org>
Tested-by: Nick Crews (Nov 1 last day) <ncrews@chromium.org>
Reviewed-by: Mathew King <mathewk@chromium.org>