overlay: Set the double_extend_pcr_issue flag for affected overlays
There are few overlays that are affected by firmware bug that extends
PCR twice in some scenarios. This affects some use cases, and to take
care of them we need a USE flag to have this information.

This is intended to be used together with CL:1942267, CL:1942227,
CL:1944147, CL:2237271 and CL:2237273

BUG=b:127321828
TEST=Manual test

Change-Id: Id76ed1544ef31642fa70453b894e07c632750d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1942949
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>