sarien: Add config definitions to public overlay
This change is part of a batch of CLs to make the Sarien overlay public.
As a first step this CL updates the public model.yaml file to include
definitions for the sarien, sarien-signed, arcada, and arcada-signed
configs. In addition, common_config is renamed to base_config to match
the private overlay.

BUG=b:138703176
TEST=compare the config.yaml files before and after these changes:
$ emerge-grunt -q chromeos-config-bsp-sarien \
	chromeos-config-bsp-sarien-private chromeos-config

Compare /build/sarien/usr/share/chromeos-config/yaml/config.yaml with
these changes to ToT and verify there are no differences.

Change-Id: I35299a7047c461bd6aa3778d1a158a0a4344adec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1913891
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Mike Wiitala <mwiitala@google.com>
Tested-by: Mike Wiitala <mwiitala@google.com>