sarien: Add signed device-names and test-label attribute to public overlay
This change is part of a batch of CLs to make the Sarien overlay public.
This CL updates the public model.yaml file to include the "test-label"
attribute previously defined in the private overlay. This CL also sets
the $device-name variable for the signed config variations in the
public overlay.

To prevent the audio attribute from being affected by the
added $device-name initializations, the $audio-device-name variable is
added and used in place of $device-name in the audio section.

BUG=b:138703176
TEST=compare the config.yaml files before and after these changes:
$ emerge-grunt -q chromeos-config-bsp-sarien \
	chromeos-config-bsp-sarien-private chromeos-config

Compare /build/sarien/usr/share/chromeos-config/yaml/config.yaml with
these changes to ToT and verify there are no differences.

Change-Id: Icc08c992b8c322c7cb234cec04c9b5948bd61269
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1913894
Tested-by: Mike Wiitala <mwiitala@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Mike Wiitala <mwiitala@google.com>