grunt: Enable pcserial for kernel
Lets compile the serial drivers into the kernel. This doesn't enable
the serial console, it just provides the drivers. This avoids the pain
of grabbing a release image, having to recompile the kernel and updating
it when trying to enable serial console manually.

BUG=none
TEST=Booted the kernel and made sure there was no serial output.
Also made sure /dev/ttyS0 was present.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I59d87104f16da6148803d7ae02fd439d9ca9fdd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1919400
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>