baseboard-trogdor: Add coreboot-private-files for TCPC firmware
This patch adds a coreboot-private-files virtual to correctly depend on
the TCPC firmware. On other boards this is done in a private overlay,
but since the TCPC firmware itself is public I don't see why this
shouldn't be public as well.

Since we don't have a coreboot-private-files virtual for Trogdor yet I'm
putting it in the public baseboard for now, but since we'll eventually
need it in the private overlays I'm separating the real ebuild contents
in a separate coreboot-private-files-baseboard-trogdor ebuild that can
later be depended on by the private overlay.

BUG=b:145723652
TEST=Built Trogdor, confirmed that TCPC firmware lands in CBFS.

Cq-Depend: chromium:1959006
Change-Id: If9c0c32b8acb413714b08abc9515bd41a101a063
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1959505
Commit-Queue: Julius Werner <jwerner@chromium.org>
Tested-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>