setzer: Enable -Os by default
This enables -Os across the board for setzer, as a part of our image
size reduction efforts.

We're trialing this on setzer and bob for the current release. If things
look good, we'll push this default globally.

It's important to note that -Os on clang is more "optimize for a mixture
of size and speed," whereas -Oz is "size at all costs (within reason)."
Hot binaries like Chrome, graphics-related things, the kernel, glibc,
etc. have already been opted out of this default.

This is part of the Cr -Os effort detailed here:
https://docs.google.com/document/d/15Z-IV7vQKrsxeeDz08nMUrJ5xKo8cIollhO-NILaNwo/edit

BUG=chromium:908661
TEST=`emerge-setzer app-admin/sudo` showed `-Os` after `-O2` in its
     build commands; setzer-release-tryjob builds, but hwtests time out
     waiting for an available device.

Cq-Depend: chromium:1970796, chromium:1968374
Change-Id: I1b7ce8fdb1146eea697f7ebb7259ce8836dab3d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1967549
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>