bob: Enable -Os by default
This enables -Os across the board for bob, as a part of our image
size reduction efforts.

We're trialing this on bob and bob for the current release. If things
look good, we'll push this default globally.

It's important to note that -Os on clang is more "optimize for a mixture
of size and speed," whereas -Oz is "size at all costs (within reason)."
Hot binaries like Chrome, graphics-related things, the kernel, glibc,
etc. have already been opted out of this default.

This is part of the Cr -Os effort detailed here:
https://docs.google.com/document/d/15Z-IV7vQKrsxeeDz08nMUrJ5xKo8cIollhO-NILaNwo/edit

BUG=chromium:908661
TEST=`emerge-bob app-admin/sudo` showed `-Os` after `-O2` in its build
     commands; bob-release-tryjob passes.

Cq-Depend: chromium:1970796, chromium:1968374, chromium:1970953
Change-Id: I785da5a22f13e0e004b908fd5efc75fac268fd5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1967771
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>