coreboot-zork: Allow adding payload
We can't easily use the coreboot build system to add a payload since
ebuilds not are permitted to clone git repos on an ad-hoc basic. So add
a feature to add seabios or U-Boot to the ROM after building.

This is controlled by USE flags.

BUG=chromium:1034094
TEST=cros_workon-zork start coreboot-zork
USE=u-boot emerge-zork coreboot-zork
$ cbfstool ~/cosarm/chroot/build/zork/firmware/trembyle/coreboot.rom print
See that fallback/payload is in there

Change-Id: I2a4c5e80d21c4d621367b32b981fcf6330e81b30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1973720
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>