audit: Filter out icf flag because the linker is BFD.
As a broader effort to reduce image size, we want to use icf=all flag
for all Chrome OS packages. Unfortunately, packages that are built using
BFD will not be able to apply this flag. So this patch is to filter out
this flag in audit for lakitu.

BUG=chromium:1022226, b:146578151
TEST=lakitu tryjob pass https://ci.chromium.org/p/chromeos/builders/general/Try/b8893594101331169568
TEST=kumo tryjob pass https://ci.chromium.org/p/chromeos/builders/general/Try/b8893594100537799968
RELEASE_NOTE=None

Change-Id: I306635eeb06512c91ffdbb848f69ece2d088e405
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1977334
Reviewed-by: Xuewei Zhang <xueweiz@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tiancong Wang <tcwang@google.com>
Commit-Queue: Tiancong Wang <tcwang@google.com>