BDW: enable crostini on all variants
samus already has crostini unconditionally enabled via this commit:

commit 9da9edee2d44 ("Always use v4.14 kernel and audio config for samus")
https://crrev.com/c/1967536

Enable it in the rest of the BDW systems.

For just one build we enable it with the kvm_transition USE flag
defined, so that we work around http://b/134764918.  As soon as this CL
lands we can remove kvm_transition so that crostini is enabled
unconditionally.

BUG=b:144308685
TEST=Verify with equery that kvm_host USE flag is set for all BDW
variants.  Manually tested on one chromebox (guado) and on one
chromebook (auron_paine), including testing with GPU acceleration turned
on in chrome flags.

Change-Id: I63c91044b93ee687b7ad5e9eafd625bc49c637a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1986938
Tested-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Ross Zwisler <zwisler@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>