u-boot: Filter out ICF flag because overlay-beaglebone is using BFD.
An earlier patch starts to use icf=all for all Chrome OS boards for
size optimizations. But there are some pacakges for boards that are
still using BFD, which doesn't recognize this flag. This patch intends
to filter out this flag for such package.

BUG=chromium:1022226,chromium:1039206
TEST=beaglebone-release-tryjob https://ci.chromium.org/p/chromeos/builders/general/Try/b8891992081941868992
TEST=beaglebone_servo-release-tryjob https://ci.chromium.org/p/chromeos/builders/general/Try/b8891992081415389648

Change-Id: Ibe8d7e6cf5d13baf96b1818aa706b5625df7239f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1988344
Commit-Queue: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>