reland BDW: have all boards use v4.14 by default
This is a reland of https://crrev.com/c/1993692.  This previous attempt
was failing in guado-accelerator, which is obsolete and has since been
removed from the CQ with these two CLs:

https://crrev.com/c/1993403
https://crrev.com/i/2397486

We're already using the new v4.14 kernel everywhere, now just make it
the default.

The only board that actually changes behavior with this change is guado,
where the new kernel has been used in guado-kernelnext.

BUG=b:129963915
TEST=Build all BDW variants.  Verify with USE flags that they are all
using the v4.14 kernel.

Change-Id: Ic57fad5cb4f916691d86ded414400499b0339c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/1997255
Tested-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Ross Zwisler <zwisler@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>