Replace Min/MaxSoftwareGain with IntrisicSensitivity
We used to use MaxSoftwareGain as an indicator of whether we want to
apply software gain instead of hardware one. After chromium:2006931
both Min and MaxSoftwareGain will be deprecated.

This change targets to keep our existing behavior for capture.
The logic is:
- IntrinsicSensitivity -2600 means we apply 20dB software gain as we
target to get -6 dBFS samples in general.
- If MaxSoftwareGain is specified >= 2000, it means we original apply
20dB sw gain from system gain, which means we need to set the
IntrinsicSensitivity to be -2600
- If MaxSoftwareGain is specified as x which is < 2000, we set the
IntrinsicSensitivity to be -600 - x following formar logic.
- If a DefaultNodeGain is specified as x with a MaxSoftwareGain defined,
it means there is an extra x/100 dB gain apply to the node beside the
system gain. We need to minus the IntrinsicSensitivity by x in order to
keep that gain.
- If a DefaultNodeGain x is specified without a MaxSoftwareGain defined,
it means we want to apply an extra x/100 dB hardware gain. Considering
the removal of a default 20 dB system gain, we need set the
DefaultNodeGain to x + 2000.

BUG=b:147080207
TEST=Build,deploy and test on DUT

Cq-Depend: chromium:2006931
Change-Id: I219fd1bc27f2a33a37d9e2ee8e2fbaecb6b1d672
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2007029
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>