BDW: enable crostini without flag on all variants
We enabled crostini in all Broadwell variants with this commit:

https://crrev.com/c/1986938

But to get through the CQ we enabled it with the kvm_transition USE flag
defined, so that we work around http://b/134764918.  Remove that flag so
that crostini is enabled unconditionally.

BUG=b:144308685
TEST=none

Change-Id: Ia16424ed7d0c3781346aa879576658814aac05a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2013998
Tested-by: Ross Zwisler <zwisler@google.com>
Auto-Submit: Ross Zwisler <zwisler@google.com>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Jacob Rasmussen <jacobraz@google.com>
Commit-Queue: Ross Zwisler <zwisler@google.com>