arc-codec-chipset-bsw: Change SW H263 and MPEG4 decoders perf values
Kernel uprev from 3.18 to 4.19 causes sw H263 and MPEG4 decoders
regressions on Ultima. It causes failures of performance tests in
CTS. This changes the sw decoders' perf values of Braswell so
that the tests pass.

The values are modified based on the new perf values range,
210-240 for android.media.cts.VideoDecoderPerfTest#testH263Goog0Perf0176x0144 and
250-280 for android.media.cts.VideoDecoderPerfTest#testMpeg4Goog0Perf0176x0144.

BUG=chromium:144525762
TEST=None

Change-Id: I3c9486eecde442a512b4e6501d9975cdc79ed442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2030347
Reviewed-by: David Staessens <dstaessens@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>