Pin monroe to mesa 19.0(-r19).
Monroe has problems with the new mesa. It also is the last Haswell and
going to be deprecated in M83. No point in upreving with it.

Notice we add the old driver in the monroe overlay so everything gets
deleted together when monroe is EOLed.

BUG=chromium:1048707
TEST=Ran monroe image on peppy.

Change-Id: I0d54e49f0c0482cf801549ff7febb9d2175c8e77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2038330
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Lepton Wu <lepton@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>