atlas: Enable or disable PSR on boot based on device model.
This is a refinement of CL:1950714 (git commit
ba76e488b42c4c961c70518304ea098a97e2c5ca) to disable PSR only on the
models commonly affected by the PSR exit flicker panel firmware bug.

This explicitly enables or disables PSR for all expected models, so in
principle this makes the aforementioned CL obsolete.  However it is
recommended to keep that CL in place alongside this, so that if this
fails or does not take effect initially, the default avoids the flicker
bug.

BUG=b:136040671,chromium:939979
TEST=Built OS image, installed on FHD and UHD DUTs.
Rebooted repeatedly with post-boot PSR setting logged.
Verified correct PSR setting for the display was in effect after every boot.
Repeated the same testing without CL:1950714.

Change-Id: I2181685ec4eaf387d328cafe6bfc843623b887b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2042797
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Nick Sanders <nsanders@chromium.org>
Commit-Queue: Matthew Blecker <matthewb@chromium.org>
Tested-by: Matthew Blecker <matthewb@chromium.org>