BDW: enable crostini without flag on chromebook variants
We enabled crostini in all Broadwell variants with this commit:

https://crrev.com/c/1986938

But to get through the CQ we enabled it with the kvm_transition USE flag
defined, so that we work around http://b/134764918.  Remove that flag so
that crostini is enabled unconditionally.

We are doing this for only the chromebook variants currently because the
transition from not having KVM enabled to having it enabled requires a
cold reboot.  For chromebooks this cold reboot is done via ectool here:

http://cs/chromeos_public/src/third_party/autotest/files/server/hosts/cros_repair.py?l=647

which was introduced by this commit:

https://crrev.com/c/1262395

Chromeboxes don't have an EC, though, so we'll have to be more creative
to be able to do the cold reboot.

BUG=b:144308685,b:134764918
TEST=none

Change-Id: Ia0669d96feb08330437a73f6ba577005d5836719
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2052600
Tested-by: Ross Zwisler <zwisler@google.com>
Auto-Submit: Ross Zwisler <zwisler@google.com>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Ross Zwisler <zwisler@google.com>