chipset-cml: Add power governor selection feature
This patch adds the feature of setting scaling_governor
to performance on AC else set it to powersave governor.
This improves performance of the device on AC.
Used udev rule to implement this functionality.

BUG=b:143941387
TEST=Added the udev rule manually and checked that udev
rule was triggered correctly. Also, manually tested the
behavior of scaling_governor by plugging in/out the AC.

Cq-Depend: chromium:2105173
Change-Id: I10580b2cd4c93cae7374f488a0b26dc1874b992c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2071554
Tested-by: Neha Kumari <neha1.kumari@intel.corp-partner.google.com>
Tested-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Mengqi Guo <mqg@chromium.org>