Zork: Replace Min/MaxSoftwareGain with IntrisicSensitivity
We used to use MaxSoftwareGain as an indicator of whether we want to
apply software gain instead of hardware one. After chromium:2006931
both Min and MaxSoftwareGain will be deprecated.

This change targets to keep our existing behavior for capture.
The logic is:
- IntrinsicSensitivity -2600 means we apply 20dB software gain as we
target to get -6 dBFS samples in general.

BUG=b:147080207
TEST=Build, deploy and test on DUT

Change-Id: I17dd435e99880e9c9920acc09fbafaa3012b23b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2084172
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>