lakitu: Fix ssh hostkey verification error
This CL does two things to fix the error:
1. google-instance-setup.service no longer generates the ssh hostkeys on
first boot since we have a script that is run before sshd.service that
does this. Instead, google-instance-setup.service will just publish the
keys found in /mnt/stateful_partition/etc/ssh to the instance guest
attributes. This change is reflected in
20190801-publish-hostkeys-from-stateful-partition.patch.
2. Makes google-instance-setup.service block on sshd.service. This is to
guarantee that the hostkeys generated by the script run before
sshd.service are present in /mnt/stateful_partition/etc/ssh when
google-instance-setup.service runs.
3. compute-image-packages relies on a config file found in /etc/default
to check whether the instance id of the VM has changed after a reboot.
This does not work on COS since /etc is stateless, so a fix for this is
made in 20190801-publish-hostkeys-from-stateful-partition.patch.

BUG=b:147559155
TEST=cos_tryjob : (https://ci.chromium.org/p/chromeos/builders/general/Try/b8886703994294750944?)
RELEASE_NOTE=Fix ssh hostkey verificaiton error

Change-Id: Icf60602907b08d214af4794784777ac89b34ce47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2088821
Commit-Queue: Dexter Rivera <riverade@chromium.org>
Tested-by: Dexter Rivera <riverade@chromium.org>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>