tast-use-flags: Add hana to use flags
Crostini tests have only recently become stable enough for the CQ.
We are enabling them across boards as we see the boards meet the
stability criteria for the CQ (20 consecutive passing runs)

Hana is meant to be disabled in the Crostini tests blacklist,
however the blacklist isn't working for hana as this use flag
is never declared.

https://source.chromium.org/chromiumos/chromiumos/codesearch/+/master:src/platform/tast/src/chromiumos/tast/cmd/local_test_runner/main.go;l=44?q=crostiniUnstableBoards&ss=chromiumos&originalUrl=https:%2F%2Fcs.chromium.org%2F

Once we see 20 consecutive passing runs of all the crostini CQ
tests on hana we will take it out of the blacklist.

BUG=chromium:1059078
TEST=./build_packages --board=hana

Change-Id: I32266d15dc4c3074a9a0789245d5ce30e6ff310e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2093956
Commit-Queue: Matthew Warton <mwarton@google.com>
Tested-by: Matthew Warton <mwarton@google.com>
Auto-Submit: Matthew Warton <mwarton@google.com>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>