Add force_breakpad USE flag
Crashpad does not work on boards with the
https://chromium-review.googlesource.com/q/Iaaac2021d7c8b68afce59856f0aaf0a156fdd43d
kernel patch. Add a USE flag so that we can disable crashpad tests on
those boards and so we can disable the crashpad experiment on those
boards.

This only affects tast tests at the moment; the experiment isn't checked
in yet.

BUG=chromium:1051354
TEST=tast ... ui.ChromeCrash* on hana, ensured crashpad tests were skipped

Change-Id: Ic91959aa2a1e9c5d3b0daebadd411a113ec6fc36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2099400
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>