lakitu: Generate kernel headers tarball
Currently, we generate the kernel headers tarball in a complicated cron
job after every build. We would like to get rid of this cron job. We can
do this by generating the kernel headers as part of the normal image
build.

We generate the kernel headers as part of the kernel install process.
The steps we take to generate the kernel headers are pretty much the
same as the steps taken by scripts/package/builddeb in the kernel tree,
which is what we currently use to build the kernel headers.

I compared the kernel headers generated at build time against the kernel
headers we generate in a post build step. The difference looks like the
following (left is current headers, right is new in-builder headers):

5772a5773,5774
> ./build
> ./build/Makefile
17981d17990
< ./scripts/unifdef
17983d17991
< ./scripts/.unifdef.cmd
18445d18452
< ./tools/objtool/sync-check.sh

The missing scripts (unifdef and sync-check.sh) appear to be used for
unnecessary things in `make bindeb-pkg` that we don't really care about.
They seem to be artifacts of the way the current process generates
headers (which is by invoking `make bindeb-pkg` and extracting
/usr/src/linux-headers-* from the resulting package).

We do not use `make bindeb-pkg` in the ebuild because the necessary
Debian packaging dependencies are not present in the cros_sdk.

I also verified that I can compile an out-of-tree kernel module with
these headers.

BUG=b:134583681
TEST=tryjob
RELEASE_NOTE=None

Change-Id: I64c966b95a0f953bb06948143728da7bb8b0dbb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2106662
Reviewed-by: Vaibhav Rustagi <vaibhavrustagi@google.com>
Commit-Queue: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>