TBR: reland: ish: stop rev'ing ISH ebuild
Since the ISH FW for Arcada (sarien) is stable and we don't want to
introduce more risk, stop automatically rev'ing the ISH ebuild with the
latest ToT changes from platform/ec. Pin the Arcada ISH FW at this point
in time.

If we need to release a bug fix in the future, we can see how ToT for
arcada_ish is doing then and we can either start taking ToT again or
make another branch.

This re-lands crrev.com/c/2116636, but as r46 instead of r43.

BUG=b:150705744
TEST=none

Change-Id: I80fd9cd39db7bc902c22451105e3959e5b30f6a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2120053
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Tested-by: LaMont Jones <lamontjones@chromium.org>