Zork: Add psp_vboot USE flag for temporary debug
We want to be able to enable this for testing, but not for
everyone.  Add a temporary USE flag.

This enables the Verstage on PSP and uses a different PSP bootloader
binary that has extra debug.

BUG=b:123887623
TEST=Build with flag and see vboot run in verstage

Cq-Depend: chrome-internal:2847921, chromium:2137638
Signed-off-by: Martin Roth <martinroth@chromium.org>
Change-Id: I25a687ecc20133908f50583e6e6e10cbde1a0691
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2137638
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Auto-Submit: Martin Roth <martinroth@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>