TO-REVERT: volteer: Add TGL PCI ID force probe
Until the force_probe condition in the i915 driver
is removed for TGL (GEN12), we will need to forcibly
probe the PCI ID for graphics to start up correctly.

After the force_probe condition is removed in the
driver, we can revert this change.

BUG=b:154037504
TEST=Build image with public overlays only.
TEST=Check src/build/images/volteer/latest/config.txt
and confirm PCI ID was added.

Change-Id: I529126a9d4f1b50438c70b61f3fef2266845ad58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2154045
Reviewed-by: Alex Levin <levinale@chromium.org>
Tested-by: Edward Baker <edward.baker@intel.com>
Commit-Queue: Alex Levin <levinale@chromium.org>