baseboard-trogdor: Simplify our ARC++ cpusets
Our overly-restrictive sets were stuffing too many things on a single
little core and causing _more_ power because it needed to bump up the
CPU frequency to handle it.  Let's simplify for now until we have a
better solution.  We'll also remove the "boost" set which nobody seems
to know anything about.

After doing this I can run (in android-sh):

  # grep "" /dev/cpuset/*/cpus
  /dev/cpuset/background/cpus:0-5
  /dev/cpuset/foreground/cpus:0-7
  /dev/cpuset/restricted/cpus:0-5
  /dev/cpuset/system-background/cpus:0-5
  /dev/cpuset/top-app/cpus:0-7

...and ARC++ still seems to work.

This also comments out the part of this file that was trying to do
something with schedtune.  It wasn't doing anything and we should
admit that.

BUG=b:157022340
TEST=See commit message

Change-Id: Id862f7178e08fa85f759c9ad0f4805052e4f2416
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2212949
Reviewed-by: Jesse Barnes <jsbarnes@google.com>
Commit-Queue: Jesse Barnes <jsbarnes@google.com>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>