arm64: There is no /dev/cpuset/foreground/boost
Nobody seems to know what this file was for and looking at systems
today it's not there anymore.  Let's delete the code that was trying
to set it.

NOTE: this _doesn't_ attempt to do a full-sale cleanup of these config
files, it just removes the one set that makes no sense.

BUG=b:157022340
TEST=None

Change-Id: Id57d913a7ddf761c55abff4ab6494bb4778939a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2231838
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>