nocturne: convert to unibuild (enable chromeos-config)
This CL converts the public overlay for nocturne to enable
chromeos-config support.  This was done entirely in an automated
manner using the cros_convert_unibuild.py script in CL:2233191.

Note: this does not mean combining the builds of nocturne with another
overlay: the build is still intended to only run on nocturne.  This is
much lower-risk than doing a complete unibuild conversion.

BUG=chromium:893725
TEST=build image for nocturne, verify all of the following:
     - /run/chromeos-config/v1 has correct contents
     - All mosys commands WAI.
     - Firmware updater working and updates firmware.
     - Fingerprint WAI.
     - Able to open Play Store, install and run Android apps.

Change-Id: Ifc7ec10e04c5337f7f94385457f329f0c68c4b91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2238669
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>