systemd: Fix the build issue arose due to meson upgrade.
Post meson upgrade from 0.44.1 to 0.54.2, `--undefined` option
expects an argument which the meson.build in systemd doesn't provide.
Due to this, it takes a random arguments that comes after the option
and the build fails.

This CL removes the `--undefined` option from the meson configuration.

BUG=b:158669562
TEST=tryjob
RELEASE_NOTE=Fix a build issue in systemd due to meson upgrade.

Change-Id: If89a00044a47f2dc23a97ade6bb0a35634600ade
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2240233
Reviewed-by: Roy Yang <royyang@google.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Commit-Queue: Vaibhav Rustagi <vaibhavrustagi@google.com>
Tested-by: Vaibhav Rustagi <vaibhavrustagi@google.com>