Enable distributed ThinLTO for auron boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for a limited set of boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
auron reference board.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: Iff45ef261e730ed9ff234e2bca8f2a30cd0bc0a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2240518
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>