Enable distributed ThinLTO for byt chipset
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for a limited set of boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
byt chipset.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: I4fff3c3f34d26e7ce0a962162ac3d0c625018acb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2251045
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>