Enable distributed ThinLTO for reef boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for a limited set of boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
reef reference board.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: I8d041ada02976772326e1b263fd3a33ec4b02128
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2252494
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>