Enable distributed ThinLTO for kunimitsu boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for a limited set of boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
kunimitsu baseboard.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: I3e93fbaf77c18841a5c96b82ab4bb987a07e59a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2254979
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>