Enable distributed ThinLTO for oak boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for a limited set of boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
oak baseboard.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: I535c94f0480e56125acdf5aa60a623a7bc0772be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2253058
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>