Enable distributed ThinLTO for jecht boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
jecht reference board.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: Ib0277ab434e952dfda7b3d4919a6ed18067e926f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2254983
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>