Enable distributed ThinLTO for kukui boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
kukui reference board.

BUG=chromium:877722
TEST=Tryjobs and hwtests for a number of affected boards

Change-Id: I8b6fc666e8fa4f7b2a30f86b9c44d97afe97676e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2258625
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>