Enable distributed ThinLTO for qc* chipsets
Relands crrev.com/c/2264886. That change was previously reverted
to reduce the load on Goma. We are now ready to re-enable.

Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
qt7180 and qt845 chipsets.

BUG=chromium:877722
TEST=Tryjobs and hwtests

Change-Id: I7611d65f529d3b3b836031c8aef81dd526b2681c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2685423
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>