Enable distributed ThinLTO for qc* chipsets
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
qt7180 and qt845 chipsets.

BUG=chromium:877722
TEST=Tryjobs and hwtests

Change-Id: I0e7e71b541ed08e8e300bbebe38e097f4ad1bdd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2264886
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>