Enable distributed ThinLTO for fizz boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
fizz baseboard.

BUG=chromium:877722
TEST=Tryjobs and hwtests

Change-Id: I045e6b13491d36a08aec4d207c1289e88177371c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2265199
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>