Enable distributed ThinLTO for glados boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more boards. This
change enables distributed ThinLTO for boards based on the
glados baseboard.

BUG=chromium:877722
TEST=Tryjobs and hwtests

Change-Id: I7b24bf11df21fbb7a5915e47f239571c9f4b98dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2268625
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>