Enable distributed ThinLTO for boards using kbl chipset
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation for
ThinLTO. This is currently enabled for most boards, and we are in the
process of enabling it for the remaining boards.  This change enables
distributed ThinLTO for boards based on the kbl chipset.

BUG=chromium:877722
TEST=Tryjobs and hwtests

Change-Id: I96196038ad91b4376f65e3ca33a9123a60690f53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2271479
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>