Enable distributed ThinLTO for boards using tgl chipset
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation for
ThinLTO. This is currently enabled for most boards, and we are in the
process of enabling it for the remaining boards. This change enables
distributed ThinLTO for boards based on the tgl chipset.

BUG=chromium:877722
TEST=Tryjobs and hwtests

Change-Id: I0e3ab3f5aa225a65063b42b59ec43bc8ce75a99d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2271481
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>