arc-codec-kukui: Disable HEVC in arc++ on kukui.
This CL disables HEVC in arc++ on kukui devices, as this is not
required on devices with detachable keyboards.

The entire arc-codec-kukui package is removed as this packages was
introduced in crrev.com/c/2266112 with the sole purpose of enabling
HEVC on kukui. Now this is no longer required we can just depend on
the default arc-codec-chipset-mt8183 package for the mt8183 chipset.

BUG=b:159412173
TEST=emerge-kukui arc-codec-chipset-mt8183 &&
     equery-kukui f arc-codec-chipset-mt8183
     checked installed xml files

Change-Id: Id34ddb974dc003ef7a587a713e7006ede8de3432
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2352461
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>