arc-codec-nocturne: Disable HEVC in arc++ on nocturne.
This CL disables HEVC in arc++ on nocturne devices, as this is not
required on devices with detachable keyboards.

The entire arc-codec-nocturne package is removed as this packages was
introduced in crrev.com/c/2265620 with the sole purpose of enabling
HEVC on nocturne. Now this is no longer required we can just depend on
the default arc-codec-chipset-kbl package for the kbl chipset.

BUG=b:159412173
TEST=emerge-nocturne arc-codec-chipset-kbl &&
     equery-nocturne f arc-codec-chipset-kbl
     checked installed xml files

Change-Id: I829505e6de0c8e399e96d31ceebd7cfd696adfe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2352462
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>