baseboard-trogdor: Delete the old schedtune config
This will be replaced with uclamp at some point and, if the
configuration still belongs here, it can be added back.  At the moment
the fact that we're trying to write to files that don't exist makes
the script exit early and the input boost isn't taking effect.

BUG=chromium:1032122, chromium:1041117
TEST=Input boost takes effect now

Change-Id: Idb843b9ce03b77df0e60c5cf5b27624974da1cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2363342
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>