kukui: get the goodix pid using model instead of chassis
"Chassis" is a concept which existed for Zerg devices (pre-unibuild
image sharing).

The model name is the equivalent in the unibuild world, and for all
unibuild devices, the chassis will be the model name in
all-caps.  (and using the chassis is deprecated)

Replace this usage with the model name.

BUG=chromium:1026435
TEST=/opt/google/touch/scripts/get_board_goodix_pid.sh on Krane

Change-Id: I79f48c1ef1c764507bb3f1775528c0a1a174b1fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2373323
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>