trogdor: Preload Qualcomm's remote processor driver
Problem: On trogdor,wifi scan and connected takes around
14~16 secs. This is much slower compared to other platforms.
This happens due to network driver's dependency on rmtfs service.

Solution: To speed up the wifi scan, in this patch we are loading
the remote processor driver early in the boot. Add a conf file which
modprobes the remote processor driver modules at start of boot-services.
This decreases the time when scan/connection starts by about 5~6 secs.

BUG=b:162465221
TEST=The build succeeds and the network drivers are getting probed
early. The scan and connect time is visibly better than before and
as per the logs it takes around 8~10 secs.

Change-Id: I6cde995fd2abf5d9a64e747f81306a7b48d16dfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2381893
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>