raspberrypi: drop unmaintained board
It's not clear anyone is using this, especially as it's for an armv6
system that we've been disabling features for, and uses an old kernel
(3.6) that we've been dropping support for.  Lets delete it so we don't
hold up other work.  If anyone cares about it, they can readd with new
packages & settings.

BUG=chromium:1117585
TEST=CQ passes

Change-Id: I1020d55eca9f4926cbd2474962da571f3dc75d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/2396535
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>